Move `FindAuthor` assoc type to `pallet_evm`
